Added command line. Initial tests
