+++ .vscode/launch.json
@@ -1,0 +1,29 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "java",
+            "name": "Current File",
+            "request": "launch",
+            "mainClass": "${file}",
+            "args": "-h"
+        },
+        {
+            "type": "java",
+            "name": "SnpSift",
+            "request": "launch",
+            "mainClass": "org.snpsift.SnpSift",
+            "projectName": "SnpSift"
+        },
+        {
+            "type": "java",
+            "name": "Zzz",
+            "request": "launch",
+            "mainClass": "org.snpsift.annotate.mem.Zzz",
+            "projectName": "SnpSift"
+        },
+    ]
+}
